+++ packages/amplify-category-auth/provider-utils/awscloudformation/triggers/PostConfirmation/add-to-group.js
@@ -1,8 +1,11 @@
-const aws = require('aws-sdk');
+const {
+  CognitoIdentityProviderClient,
+  AdminAddUserToGroupCommand,
+  GetGroupCommand,
+  CreateGroupCommand,
+} = require('@aws-sdk/client-cognito-identity-provider');
 
-const cognitoIdentityServiceProvider = new aws.CognitoIdentityServiceProvider({
-  apiVersion: '2016-04-18',
-});
+const cognitoIdentityServiceProvider = new CognitoIdentityProviderClient({});
 
 /**
  * @type {import('@types/aws-lambda').PostConfirmationTriggerHandler}
@@ -21,14 +24,14 @@ exports.handler = async (event) => {
    * Check if the group exists; if it doesn't, create it.
    */
   try {
-    await cognitoIdentityServiceProvider.getGroup(groupParams).promise();
+    await cognitoIdentityServiceProvider.send(new GetGroupCommand(groupParams));
   } catch (e) {
-    await cognitoIdentityServiceProvider.createGroup(groupParams).promise();
+    await cognitoIdentityServiceProvider.send(new CreateGroupCommand(groupParams));
   }
   /**
    * Then, add the user to the group.
    */
-  await cognitoIdentityServiceProvider.adminAddUserToGroup(addUserParams).promise();
+  await cognitoIdentityServiceProvider.send(new AdminAddUserToGroupCommand(addUserParams));
 
   return event;
 };