+++ packages/amplify-category-function/src/provider-utils/awscloudformation/utils/lambda-layer-cloudformation-template.ts
@@ -1,4 +1,4 @@
-import { AmplifyError, stateManager } from '@aws-amplify/amplify-cli-core';
+import { stateManager } from '@aws-amplify/amplify-cli-core';
 import { DeletionPolicy, Fn, IntrinsicFunction, Refs } from 'cloudform-types';
 import Lambda from 'cloudform-types/types/lambda';
 import _ from 'lodash';
@@ -30,12 +30,7 @@ export function generateLayerCfnObj(isNewVersion: boolean, parameters: LayerPara
     layerCloudState.latestVersionLogicalId = logicalName;
     versionList.unshift({ LogicalName: logicalName, legacyLayer: false });
   } else {
-    logicalName = _.first(versionList)?.LogicalName;
-    if (!logicalName) {
-      throw new AmplifyError('LambdaLayerNotFoundError', {
-        message: 'No versions were found for the Lambda Layer. Were they deleted on the AWS Lambda Console?',
-      });
-    }
+    logicalName = _.first(versionList).LogicalName;
   }
 
   const outputObj = {