+++ packages/amplify-category-geo/resources/custom-place-index-resource-handler.js
@@ -1,10 +1,10 @@
 const response = require('cfn-response');
-const aws = require('aws-sdk');
+const { LocationClient, CreatePlaceIndexCommand, DeletePlaceIndexCommand, UpdatePlaceIndexCommand } = require('@aws-sdk/client-location');
 exports.handler = async function (event, context) {
   try {
     console.log('REQUEST RECEIVED:' + JSON.stringify(event));
     const pricingPlan = 'RequestBasedUsage';
-    if (event.RequestType == 'Create') {
+    if (event.RequestType === 'Create') {
       const params = {
         IndexName: event.ResourceProperties.indexName,
         DataSource: event.ResourceProperties.dataSource,
@@ -13,8 +13,8 @@ exports.handler = async function (event, context) {
         },
         PricingPlan: pricingPlan,
       };
-      const locationClient = new aws.Location({ apiVersion: '2020-11-19', region: event.ResourceProperties.region });
-      const res = await locationClient.createPlaceIndex(params).promise();
+      const locationClient = new LocationClient({ region: event.ResourceProperties.region });
+      const res = await locationClient.send(new CreatePlaceIndexCommand(params));
       console.log('create resource response data' + JSON.stringify(res));
       if (res.IndexName && res.IndexArn) {
         event.PhysicalResourceId = res.IndexName;
@@ -23,7 +23,7 @@ exports.handler = async function (event, context) {
         await response.send(event, context, response.FAILED, res, params.IndexName);
       }
     }
-    if (event.RequestType == 'Update') {
+    if (event.RequestType === 'Update') {
       const params = {
         IndexName: event.ResourceProperties.indexName,
         DataSourceConfiguration: {
@@ -31,8 +31,8 @@ exports.handler = async function (event, context) {
         },
         PricingPlan: pricingPlan,
       };
-      const locationClient = new aws.Location({ apiVersion: '2020-11-19', region: event.ResourceProperties.region });
-      const res = await locationClient.updatePlaceIndex(params).promise();
+      const locationClient = new LocationClient({ region: event.ResourceProperties.region });
+      const res = await locationClient.send(new UpdatePlaceIndexCommand(params));
       console.log('update resource response data' + JSON.stringify(res));
       if (res.IndexName && res.IndexArn) {
         event.PhysicalResourceId = res.IndexName;
@@ -41,12 +41,12 @@ exports.handler = async function (event, context) {
         await response.send(event, context, response.FAILED, res, params.IndexName);
       }
     }
-    if (event.RequestType == 'Delete') {
+    if (event.RequestType === 'Delete') {
       const params = {
         IndexName: event.ResourceProperties.indexName,
       };
-      const locationClient = new aws.Location({ apiVersion: '2020-11-19', region: event.ResourceProperties.region });
-      const res = await locationClient.deletePlaceIndex(params).promise();
+      const locationClient = new LocationClient({ region: event.ResourceProperties.region });
+      const res = await locationClient.send(new DeletePlaceIndexCommand(params));
       event.PhysicalResourceId = event.ResourceProperties.indexName;
       console.log('delete resource response data' + JSON.stringify(res));
       await response.send(event, context, response.SUCCESS, res, params.IndexName);