+++ packages/amplify-e2e-core/src/utils/pinpoint.ts
@@ -20,42 +20,17 @@ const settings = {
   pinpointResourceName: 'testpinpoint',
 };
 
-const defaultPinpointRegion = 'us-east-1';
-const serviceRegionMap = {
-  'us-east-1': 'us-east-1',
-  'us-east-2': 'us-east-1',
-  'sa-east-1': 'us-east-1',
-  'ca-central-1': 'ca-central-1',
-  'us-west-1': 'us-west-2',
-  'us-west-2': 'us-west-2',
-  'cn-north-1': 'us-west-2',
-  'cn-northwest-1': 'us-west-2',
-  'ap-south-1': 'ap-south-1',
-  'ap-northeast-3': 'us-west-2',
-  'ap-northeast-2': 'ap-northeast-2',
-  'ap-southeast-1': 'ap-southeast-1',
-  'ap-southeast-2': 'ap-southeast-2',
-  'ap-northeast-1': 'ap-northeast-1',
-  'eu-central-1': 'eu-central-1',
-  'eu-north-1': 'eu-central-1',
-  'eu-south-1': 'eu-central-1',
-  'eu-west-1': 'eu-west-1',
-  'eu-west-2': 'eu-west-2',
-  'eu-west-3': 'eu-west-1',
-  'me-south-1': 'ap-south-1',
-};
-
 /**
  * checks to see if the pinpoint app exists
  */
-export async function pinpointAppExist(pinpointProjectId: string): Promise<boolean> {
+export async function pinpointAppExist(pinpointProjectId: string, region: string): Promise<boolean> {
   let result = false;
 
   const pinpointClient = new Pinpoint({
     accessKeyId: settings.accessKeyId,
     secretAccessKey: settings.secretAccessKey,
     sessionToken: settings.sessionToken,
-    region: _.get(serviceRegionMap, settings.region, defaultPinpointRegion),
+    region,
   });
 
   try {
@@ -101,7 +76,7 @@ export function initProjectForPinpoint(cwd: string): Promise<void> {
       .wait('Choose your default editor:')
       .sendLine(settings.editor)
       .wait("Choose the type of app that you're building")
-      .sendLine(settings.appType)
+      .sendLine('javascript')
       .wait('What javascript framework are you using')
       .sendLine(settings.framework)
       .wait('Source Directory Path:')