+++ packages/amplify-e2e-tests/src/__tests__/export-pull-c.test.ts
@@ -5,7 +5,6 @@ import {
   addApiWithoutSchema,
   addAuthWithMaxOptions,
   addConvert,
-  addDEVHosting,
   addS3StorageWithIdpAuth,
   amplifyPush,
   amplifyPushWithoutCodegen,
@@ -64,7 +63,6 @@ describe('amplify export pull c', () => {
   const AddandPushCategories = async (frontend?: string): Promise<void> => {
     await addAuthWithMaxOptions(projRoot, { frontend });
     await addApiWithoutSchema(projRoot, { transformerVersion: 1 });
-    await addDEVHosting(projRoot);
     await addS3StorageWithIdpAuth(projRoot);
     await addConvert(projRoot);
     if (frontend === 'flutter') {