+++ packages/amplify-migration-tests/src/migration-helpers/init.ts
@@ -43,7 +43,7 @@ export function initJSProjectWithProfileV4_28_2(
       .wait('Choose your default editor:')
       .sendLine(s.editor)
       .wait("Choose the type of app that you're building")
-      .sendLine(s.appType)
+      .sendCarriageReturn()
       .wait('What javascript framework are you using')
       .sendLine(s.framework)
       .wait('Source Directory Path:')
@@ -97,7 +97,7 @@ export function initJSProjectWithProfileV4_52_0(
       .wait('Choose your default editor:')
       .sendLine(s.editor)
       .wait("Choose the type of app that you're building")
-      .sendLine(s.appType)
+      .sendCarriageReturn()
       .wait('What javascript framework are you using')
       .sendLine(s.framework)
       .wait('Source Directory Path:')
@@ -125,6 +125,48 @@ export function initJSProjectWithProfileV4_52_0(
 }
 
 export function initAndroidProjectWithProfile(cwd: string, settings: Record<string, unknown>): Promise<void> {
+  const s = { ...defaultSettings, ...settings };
+
+  addCircleCITags(cwd);
+
+  return new Promise((resolve, reject) => {
+    spawn(getCLIPath(), ['init'], {
+      cwd,
+      stripColors: true,
+      env: {
+        CLI_DEV_INTERNAL_DISABLE_AMPLIFY_APP_CREATION: '1',
+      },
+    })
+      .wait('Enter a name for the project')
+      .sendLine(s.name)
+      .wait('Initialize the project with the above configuration?')
+      .sendConfirmNo()
+      .wait('Enter a name for the environment')
+      .sendLine(s.envName)
+      .wait('Choose your default editor:')
+      .sendLine(s.editor)
+      .wait("Choose the type of app that you're building")
+      .sendLine('android')
+      .wait('Where is your Res directory')
+      .sendCarriageReturn()
+      .wait('Select the authentication method you want to use:')
+      .sendCarriageReturn()
+      .wait('Please choose the profile you want to use')
+      .sendLine(s.profileName)
+      .wait(/Try "amplify add api" to create a backend API and then "amplify (push|publish)" to deploy everything/)
+      .run((err: Error) => {
+        if (!err) {
+          addCircleCITags(cwd);
+
+          resolve();
+        } else {
+          reject(err);
+        }
+      });
+  });
+}
+
+export function initAndroidProjectWithProfileInquirer(cwd: string, settings: Record<string, unknown>): Promise<void> {
   const s = { ...defaultSettings, ...settings };
 
   addCircleCITags(cwd);