+++ packages/amplify-console-integration-tests/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.7.1](https://github.com/aws-amplify/amplify-console-integration-tests/compare/@aws-amplify/amplify-console-integration-tests@2.7.0...@aws-amplify/amplify-console-integration-tests@2.7.1) (2023-04-29)
+
+**Note:** Version bump only for package @aws-amplify/amplify-console-integration-tests
+
+
+
+
+
+# [2.7.0](https://github.com/aws-amplify/amplify-console-integration-tests/compare/@aws-amplify/amplify-console-integration-tests@2.6.11...@aws-amplify/amplify-console-integration-tests@2.7.0) (2023-04-24)
+
+
+### Features
+
+* **amplify-util-uibuilder:** support StorageField ([#12386](https://github.com/aws-amplify/amplify-console-integration-tests/issues/12386)) ([90fe1e4](https://github.com/aws-amplify/amplify-console-integration-tests/commit/90fe1e4574c52eb4cacd110d50310f55bf913214))
+
+
+
+
+
+## [2.6.11](https://github.com/aws-amplify/amplify-console-integration-tests/compare/@aws-amplify/amplify-console-integration-tests@2.6.9...@aws-amplify/amplify-console-integration-tests@2.6.11) (2023-04-14)
+
+**Note:** Version bump only for package @aws-amplify/amplify-console-integration-tests
+
+
+
+
+
 ## [2.6.9](https://github.com/aws-amplify/amplify-console-integration-tests/compare/@aws-amplify/amplify-console-integration-tests@2.6.9-rc.e9e0d157d.0...@aws-amplify/amplify-console-integration-tests@2.6.9) (2023-03-23)
 
 **Note:** Version bump only for package @aws-amplify/amplify-console-integration-tests