+++ packages/amplify-dynamodb-simulator/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.6.0](https://github.com/aws-amplify/amplify-cli/compare/amplify-dynamodb-simulator@2.5.12...amplify-dynamodb-simulator@2.6.0) (2023-04-24)
+
+
+### Features
+
+* **amplify-util-uibuilder:** support StorageField ([#12386](https://github.com/aws-amplify/amplify-cli/issues/12386)) ([90fe1e4](https://github.com/aws-amplify/amplify-cli/commit/90fe1e4574c52eb4cacd110d50310f55bf913214))
+
+
+
+
+
+## [2.5.12](https://github.com/aws-amplify/amplify-cli/compare/amplify-dynamodb-simulator@2.5.10...amplify-dynamodb-simulator@2.5.12) (2023-04-14)
+
+**Note:** Version bump only for package amplify-dynamodb-simulator
+
+
+
+
+
 ## [2.5.10](https://github.com/aws-amplify/amplify-cli/compare/amplify-dynamodb-simulator@2.5.10-rc.e9e0d157d.0...amplify-dynamodb-simulator@2.5.10) (2023-03-23)
 
 **Note:** Version bump only for package amplify-dynamodb-simulator