+++ packages/amplify-util-import/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.5.0](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-util-import@2.4.2...@aws-amplify/amplify-util-import@2.5.0) (2023-04-24)
+
+
+### Features
+
+* **amplify-util-uibuilder:** support StorageField ([#12386](https://github.com/aws-amplify/amplify-cli/issues/12386)) ([90fe1e4](https://github.com/aws-amplify/amplify-cli/commit/90fe1e4574c52eb4cacd110d50310f55bf913214))
+
+
+
+
+
+## [2.4.2](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-util-import@2.4.0...@aws-amplify/amplify-util-import@2.4.2) (2023-04-14)
+
+**Note:** Version bump only for package @aws-amplify/amplify-util-import
+
+
+
+
+
 # 2.4.0 (2023-02-27)